[FLINK-35749] extract KafkaWriterTestBase and add fault tolerance test class
